Wired up record rate commands into controller: - The hexapod data recorder state now contains an element for the record rate. - The controller initialization command not loads the record rate when the controller starts. - Added setter and getter function to the controller class for the record rate. - Changed the record rate data type in the commands to int, as that's what the PI DLL funciton actually takes/returns

+++ source/application/AMPIC887DriverServer/GCS2Commands/AMGCS2GetRecordRateCommand.h
@@ -25,7 +25,7 @@ public:
 	  * The retrieved record rate for the controller, in controller cycles (20us).
 	  * If the command has not yet been run successfully this will return 0.
 	  */
-	double recordRate() const;
+	int recordRate() const;
 
 protected:
 
@@ -34,7 +34,7 @@ protected:
 	  */
 	virtual bool runImplementation();
 
-	double recordRate_;
+	int recordRate_;
 };
 
 #endif // AMGCS2GETRECORDRATECOMMAND_H
